% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_precision}
\alias{extract_precision}
\title{Extract precision matrices from an ssjgl fit}
\usage{
extract_precision(fit, v0_index = NULL)
}
\arguments{
\item{fit}{An object of class \code{ssjgl}.}

\item{v0_index}{Integer index into the v0 ladder. Default \code{NULL}
uses the last step.}
}
\value{
A list of K precision matrices (p x p).
}
\description{
Extract precision matrices from an ssjgl fit
}
