% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-eval.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot ROC curve}
\usage{
plot_roc(roc_obj, main = "ROC Curve", ...)
}
\arguments{
\item{roc_obj}{A list returned by \code{\link{roc_auc}}.}

\item{main}{Title for the plot. Default \code{"ROC Curve"}.}

\item{...}{Additional arguments passed to \code{\link[graphics]{plot}}.}
}
\value{
Invisible NULL. Called for side effect (plot).
}
\description{
Plots the ROC curve from a \code{\link{roc_auc}} result.
}
\examples{
truth <- matrix(c(0,1,0,0, 1,0,1,0, 0,1,0,1, 0,0,1,0), 4, 4)
scores <- matrix(c(0,.9,.1,.05, .9,0,.8,.1, .1,.8,0,.7, .05,.1,.7,0), 4, 4)
roc <- roc_auc(scores, truth)
plot_roc(roc)
}
