% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ssjgl}
\alias{print.ssjgl}
\title{Print an ssjgl object}
\usage{
\method{print}{ssjgl}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ssjgl}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisible \code{x}.
}
\description{
Print an ssjgl object
}
