% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.ssjgl}
\alias{print.summary.ssjgl}
\title{Print summary of ssjgl}
\usage{
\method{print}{summary.ssjgl}(x, ...)
}
\arguments{
\item{x}{A \code{summary.ssjgl} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print summary of ssjgl
}
