% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdat}
\alias{simdat}
\title{Simulated Network Data}
\format{
\subsection{\code{simdat}}{

A list with 3 matrices each with 150 rows and 50 columns:
\describe{
\item{V1:V100}{Data generated from multivariate normal with prespecified
block structure for each of the 3 groups}
...
}
}
}
\source{
data-raw/simdat.R
}
\usage{
simdat
}
\description{
A simulated dataset for 3 groups with slightly different block structures.
}
\keyword{datasets}
