% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikeyglass-package.R
\docType{package}
\name{spikeyglass-package}
\alias{spikeyglass}
\alias{spikeyglass-package}
\title{spikeyglass: Bayesian Joint Spike-and-slab Graphical Lasso}
\description{
Package implementing the method Bayesian joint spike-and-slab graphical lasso to infer multiple Gaussian graphical models in R. Bayesian treatments of group graphical lasso and fused graphical lasso.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mljaniczek.github.io/spikeyglass/}
}

}
\author{
\strong{Maintainer}: Margaret Janiczek \email{mjaniczek@umass.edu}

Authors:
\itemize{
  \item Zehang Richard Li \email{lizehang@gmail.com}
}

}
\keyword{internal}
