% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.ssjgl}
\alias{summary.ssjgl}
\title{Summarize an ssjgl object}
\usage{
\method{summary}{ssjgl}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ssjgl}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A list of class \code{summary.ssjgl} with summary information.
}
\description{
Summarize an ssjgl object
}
