% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.ssjgl}
\alias{coef.ssjgl}
\title{Extract precision matrices from an ssjgl fit}
\usage{
\method{coef}{ssjgl}(object, v0_index = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{ssjgl}.}

\item{v0_index}{Integer index into the v0 ladder. Default \code{NULL}
uses the last step (most sparse).}

\item{...}{Additional arguments (ignored).}
}
\value{
A list of K precision matrices (p x p).
}
\description{
Extract precision matrices from an ssjgl fit
}
