% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_precision}
\alias{extract_precision}
\title{Extract precision matrices from an ssjgl fit}
\usage{
extract_precision(fit, v0_index = NULL)
}
\arguments{
\item{fit}{An object of class \code{ssjgl}.}

\item{v0_index}{Integer index into the v0 ladder. Default \code{NULL}
uses the last step.}
}
\value{
A list of K precision matrices (p x p).
}
\description{
Equivalent to \code{coef(fit)}.
}
\examples{
sim <- simulate_ssjgl_data(K = 2, p = 10, n = 50, seed = 1)
fit <- ssjgl(sim$data_list, penalty = "fused",
             lambda0 = 1, lambda1 = 0.5, lambda2 = 0.5,
             v0s = 0.01, maxitr.em = 10, impute = FALSE)
theta <- extract_precision(fit)
dim(theta[[1]])
}
