% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_probabilities}
\alias{extract_probabilities}
\title{Extract edge inclusion probabilities from an ssjgl fit}
\usage{
extract_probabilities(fit, v0_index = NULL)
}
\arguments{
\item{fit}{An object of class \code{ssjgl}.}

\item{v0_index}{Integer index into the v0 ladder. Default \code{NULL}
uses the last step.}
}
\value{
A list with elements:
\describe{
\item{prob1}{p x p matrix of edge inclusion probabilities P(delta=1).}
\item{prob2}{p x p matrix of non-similarity probabilities P(xi=1),
or NULL if not doubly spike-and-slab.}
}
}
\description{
Extract edge inclusion probabilities from an ssjgl fit
}
