% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdiffmetric.R
\name{getdiffmetric}
\alias{getdiffmetric}
\title{getdiffmetric}
\usage{
getdiffmetric(est, truth, graph, tol)
}
\arguments{
\item{est}{?}

\item{truth}{?}

\item{graph}{?}

\item{tol}{?}
}
\value{
Summary of true positive, false positives, true negatives etc
}
\description{
getdiffmetric
}
