% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmetric.R
\name{getmetric}
\alias{getmetric}
\title{getmetric}
\usage{
getmetric(est, truth, graph)
}
\arguments{
\item{est}{?}

\item{truth}{?}

\item{graph}{?}
}
\value{
Summary of true positive, false positives, true negatives etc
}
\description{
getmetric
}
