% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_path.R
\name{plot_path}
\alias{plot_path}
\title{Plot the solution path}
\usage{
plot_path(
  v0s,
  obj,
  G,
  thres = 0.5,
  normalize = FALSE,
  xlab = "",
  ylab = "",
  main = "",
  par = c(1, 2),
  ylim = NULL,
  reverse = FALSE,
  position = "bottomright",
  only = NULL,
  color0 = "gray70",
  color1 = "darkblue",
  which.top = c(0, 1)[2],
  cex = 0.5,
  cex.xlab = 1,
  cex.ylab = 1,
  vline = NULL,
  ...
)
}
\arguments{
\item{v0s}{list of v0 input}

\item{obj}{object created using ssjgl}
}
\value{
plot
}
\description{
Plots the solution paths from ssjgl().
}
