% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_path.R
\name{plot_path}
\alias{plot_path}
\title{Plot the solution path}
\usage{
plot_path(
  v0s,
  obj,
  thres = 0.5,
  normalize = FALSE,
  xlab = "",
  ylab = "",
  main = "",
  par = c(1, 2),
  ylim = NULL,
  reverse = FALSE,
  position = "bottomright",
  ...
)
}
\arguments{
\item{v0s}{list of v0 input}

\item{obj}{object created using ssjgl}

\item{thres}{default threshold for probability of edge inclusion}

\item{normalize}{True or False (default False) on if normalized covariance should be returned}

\item{xlab}{String for xlab. Default empty.}

\item{ylab}{string for y-axis label. Default empty.}

\item{main}{string for main title. Default empty.}

\item{par}{default c(1,2) to list layout of output plots}

\item{ylim}{default NULL and will calculate from range}

\item{reverse}{Default FALSE. If want to reverse direction of plot.}

\item{position}{Default "bottomright". For position of legend.}

\item{...}{other parameters passed in to base plot}
}
\value{
plot
}
\description{
Plots the solution paths from ssjgl().
}
