% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssjgl.R
\name{ssjgl}
\alias{ssjgl}
\title{ssjgl}
\usage{
ssjgl(
  Y,
  penalty = "fused",
  lambda0,
  lambda1,
  lambda2,
  v1 = 1,
  v0s = seq(1e-04, 0.01, len = 10),
  doubly = FALSE,
  rho = 1,
  a = 1,
  b = 1,
  maxitr.em = 500,
  tol.em = 1e-04,
  maxitr.jgl = 500,
  tol.jgl = 1e-05,
  warm = NULL,
  warm.connected = NULL,
  truncate = 1e-05,
  normalize = FALSE,
  c = 0.1,
  impute = TRUE
)
}
\arguments{
\item{Y}{List of k data matrices}

\item{penalty}{Either "fused" or "group"}

\item{lambda0}{scalar for penalization of the diagonals}

\item{lambda1}{either constant or matrix of values to search over}

\item{lambda2}{either constant or matrix of values to search over}

\item{v1}{edgewise penalties}

\item{v0s}{edgewise penalties}

\item{doubly}{True or False}

\item{rho}{default as 1}

\item{a}{initializing parameters}

\item{b}{initializing parameters}

\item{maxitr.em}{max iterations of EM algorithm. Default 500}

\item{tol.em}{default 1e-4 ADD MORE}

\item{maxitr.jgl}{max iterations for JGL. default 500}

\item{tol.jgl}{default 1e-4 ADD MORE}

\item{warm}{default NULL. warming parameter for M step}

\item{warm.connected}{parameter for M-step}

\item{truncate}{cutoff to truncate default 1e-5}

\item{normalize}{True or False}

\item{c}{constant. Default 0.1}

\item{impute}{true or false}
}
\value{
similar output to JGL
}
\description{
ssjgl
}
