% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.ssjgl}
\alias{summary.ssjgl}
\title{Summarize an ssjgl object}
\usage{
\method{summary}{ssjgl}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ssjgl}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A list of class \code{summary.ssjgl} with summary information.
}
\description{
Summarize an ssjgl object
}
\examples{
sim <- simulate_ssjgl_data(K = 2, p = 10, n = 50, seed = 1)
fit <- ssjgl(sim$data_list, penalty = "fused",
             lambda0 = 1, lambda1 = 0.5, lambda2 = 0.5,
             v0s = 0.01, maxitr.em = 10, impute = FALSE)
summary(fit)
}
